Add testing of :defined prior to call to super()
Per the conversation here [1], there is a desire to add more
testing of the :defined pseudo state, prior to the call to super().
Note that `this` is not accessible prior to super(), so the
instance itself is used. Also note that :defined already does not
match anywhere inside the constructor, for upgrades.

[1] https://github.com/whatwg/dom/pull/894#discussion_r495092027

Bug: 1042130
Change-Id: I2372900981247ea5624e737d2597d004398de477